dac pins/spi init with bus manager seems to work

+++ src/hardware/driver/dac.rs
@@ -2,14 +2,14 @@ use embedded_hal::blocking::spi::Transfer;
 ///! Driver DAC11001 driver
 use stm32h7xx_hal::gpio;
 
-pub struct Dac<SPI: Transfer<u32>> {
+pub struct Dac<SPI: Transfer<u8>> {
     spi: SPI,
     sync_n: gpio::ErasedPin<gpio::Output>,
 }
 
 impl<SPI> Dac<SPI>
 where
-    SPI: Transfer<u32>,
+    SPI: Transfer<u8>,
 {
     pub fn new(spi: SPI, sync_n: gpio::ErasedPin<gpio::Output>) -> Self {
         Dac { spi, sync_n }
